+++ back/src/main/java/com/seguridadlimite/models/aprendiz/domain/AprendizSinRelaciones.java
@@ -1,6 +1,12 @@
 package com.seguridadlimite.models.aprendiz.domain;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Basic;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
@@ -18,7 +24,7 @@ public class AprendizSinRelaciones implements Serializable {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Basic(optional = false)
   @Column(name = "id")
-  private Long id;
+  private Integer id;
 
   @Basic(optional = false)
   @NotNull
