+++ back/src/main/java/com/seguridadlimite/models/aprendiz/infrastructure/repository/AprendizRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AprendizRepository extends JpaRepository<Aprendiz, Long> {
+public interface AprendizRepository extends JpaRepository<Aprendiz, Integer> {
     
     @Query("SELECT a FROM Aprendiz a JOIN FETCH a.trabajador WHERE a.idPermiso = :idPermiso")
     List<Aprendiz> findByIdPermisoWithTrabajador(@Param("idPermiso") Integer idPermiso);
